backend/requirements.txt
# FastAPI and web framework
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.0,<5.0.0

# HTTP client
httpx>=0.27.0,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0

# Redis and task queue
redis>=5.0.0,<6.0.0
rq>=1.16.0,<2.0.0

# Validation and serialization
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# ImageMagick Python bindings
Wand>=0.6.13,<1.0.0

# File handling
aiofiles>=23.2.0,<24.0.0
python-magic>=0.4.27,<1.0.0
Pillow>=10.2.0,<12.0.0

# AI Background Removal
rembg>=2.0.50,<3.0.0
onnxruntime>=1.16.0,<2.0.0
pymatting>=1.1.8,<2.0.0
scipy>=1.11.0,<2.0.0

# Security
slowapi>=0.1.9,<1.0.0
limits>=3.7.0,<4.0.0

# Testing (compatible versions)
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<0.24.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
orjson>=3.9.0,<4.0.0
